feat: standardize headlines and reorganize IA hierarchy
