+++ src/app/api/passive/settings/route.ts
