+++ src/app/api/passive/snapshots/[date]/label/route.ts
